src/model_deployment/requirements.txt
numba
librosa
matplotlib
opencv-python
opencv-contrib-python
Pillow
tqdm
yacs
future
numpy
scipy
scikit-learn
tensorboard
google-cloud-storage
google-cloud-aiplatform
wandb
soundfile
chardet
charset_normalizer
protobuf
moviepy
google
pandas
